% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateCrisprNSObject}
\alias{CreateCrisprNSObject}
\title{Create a Crispr Negative Selection Object}
\usage{
CreateCrisprNSObject(
  counts,
  project = "CrisprNegativeSelection",
  library.metadata,
  meta.data,
  Group,
  Replicate,
  Stage
)
}
\arguments{
\item{Stage}{}
}
\value{

}
\description{
Create a Crispr Negative Selection Object
}
