% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{CrisprNS-class}
\alias{CrisprNS-class}
\alias{CrisprNS}
\title{Crispr Negative Selection Object}
\value{

}
\description{
Crispr Negative Selection Object
}
\section{Slots}{

\describe{
\item{\code{project}}{character.}

\item{\code{raw.data}}{data.frame.}

\item{\code{data}}{data.frame.}

\item{\code{fc}}{data.frame.}

\item{\code{compare}}{character.}

\item{\code{meta.data}}{data.frame.}

\item{\code{replicate}}{character.}

\item{\code{version}}{ANY.}
}}

