% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{DrawFC}
\alias{DrawFC}
\title{Draw scatter plot}
\usage{
DrawFC(
  object,
  Replicate = NULL,
  xSample = "xsample",
  ySample = "ysample",
  limitation = c(-4, 4),
  breaks = c(-4, -3, -2, -1, 0, 1, 2, 3, 4)
)
}
\arguments{
\item{breaks}{}
}
\value{

}
\description{
Draw scatter plot
}
