% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{NormalizeData}
\alias{NormalizeData}
\title{Normalize raw data counts}
\usage{
NormalizeData(object, initialStage = "Day0", controlsgRNA = "nontargeting")
}
\arguments{
\item{controlsgRNA}{}
}
\value{

}
\description{
Normalize raw data counts
}
