% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{NormalizeData}
\alias{NormalizeData}
\title{Normalize raw data counts}
\usage{
NormalizeData(object, initialStage = NULL, controlsgRNA = "nontargeting")
}
\arguments{
\item{controlsgRNA}{}
}
\value{
CrisprNS object with normalized, logFC, logMean slots
}
\description{
Normalize raw data counts
}
\examples{
test <- NormalizeData(test,initialStage = "Day0",controlsgRNA = "nontargeting")
}
